muxink: Move `TestingStream` to its own submodule
